Jack transport. Missing option to turn on/off, plus requesting jack transport to play/stop/etc instead of calling the equivilant functions in Radium directly when program is configured to use jack transport.

+++ api/api_seqtracks.c
@@ -558,7 +558,7 @@ void setSeqtempoLength(double end_time, bool do_shrink){
 double getSeqtempoLength(void){
   return TEMPOAUTOMATION_get_length();
 }
-double getSeqtempoAbsabstime(double abstime){
+int64_t getSeqtempoAbsabstime(double abstime){
   return TEMPOAUTOMATION_get_absabstime(abstime);
 }
